LU-7101 lnet: per NI map-on-demand value
Enables support of different map-on-demand values per NI.  This is
required to support OPA coexistence with MLX5 cards.  MLX5 does not
support FMR, which is enabled via map-on-demand.  However OPA's
performance is greatly enahanced when FMR is enabled.  In order
to enable coexistence of both of these two types of cards we
need to be able to set different map-on-demand values for both
NIs.

Lustre-change: http://review.whamcloud.com/16367
Lustre-commit: 94d1ce46562bd36c8959a9458cacabb7f6df681f

This patch also lays the ground work for other per NI tunables to
be added in future patches.

Signed-off-by: Doug Oucharek <doug.s.oucharek@intel.com>
Change-Id: If7b7fe4323ed19d55398f1442cd33daf780e1b78
Reviewed-on: http://review.whamcloud.com/20073
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Amir Shehata <amir.shehata@intel.com>
Reviewed-by: Liang Zhen <liang.zhen@intel.com>
Reviewed-by: John L. Hammond <john.hammond@intel.com>