LU-6679 ldlm: do not send blocking ast for group locks
Group locks are acquired and released manually on client so
it doesn't make sense to send blocking AST to client when
there is an incompatible lock enqueued.

Currently client will set CBPENDING on the group lock when
it receives a blocking AST. Having the CBPENDING flag set
will make ldlm_lock_match to fail and there will be two
group locks granted on the same resource on the client.
--
The above is the upstream description.

This bug was encountered running the parallel_grouplock
test from the parallel-scale part of the Lustre test suite.

The fix was tested by Justin Miller and confirmed to
resolve the bug.  Note this is a SERVER SIDE fix.

Bugs Affected: 837384
Bug Action: RESOLVE