LU-7890 lov: Ensure correct operation for large object sizes
If a backing filesystem (ZFS) returns that it supports very large
(LLONG_MAX) object sizes, that should be correctly supported.  This
fixes the check for unitialized stripe_maxbytes in
lsm_unpackmd_common(), so that ZFS can return LLONG_MAX and it will be
okay.  This issue is excersized by writing to or past the 2TB boundry
of a singly stripped file.  A regression test is added to check for
this.

Lustre-change: http://review.whamcloud.com/19066
Lustre-commit: 3ddcf5b4a13851805f15b39aa1e95aee9b291132

Signed-off-by: Nathaniel Clark <nathaniel.l.clark@intel.com>
Change-Id: Ia5d9988dc4bdd4da08ff46ed1c1eda8cf2876e69
Reviewed-on: http://review.whamcloud.com/20142
Reviewed-by: Andreas Dilger <andreas.dilger@intel.com>
Tested-by: Jenkins
Reviewed-by: Jinshan Xiong <jinshan.xiong@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: John L. Hammond <john.hammond@intel.com>