Use binary tree in ldlm_lock_match for extent locks
Currently, ldlm_lock_match uses the search_queue code to look for
matching locks.  That code does a linear search of the lock list for the
resource.  That was fine prior to lockahead, when it was almost
impossible for a client (and this is client only code) to have more
than a handful of locks on a single resource (Almost always < 5).

With lockahead, it's easy to have thousands of locks on one resource.
When cleaning up locks for ldlm_drop_caches, we end up doing a linear
search of this list for every lock on the list.  So, complexity n^2.

This is done to see if another lock also covers a page under the lock
we are destroying, which is possible for read locks.  (In that case,
we should keep the page rather than free it.)

This was taking an unacceptable amount of time and causing heartbeat
losses because of it.

This can be avoided by using the red black binary tree which stores the
extents of granted locks, reducing the complexity of an individual
search from 'n' to 'log n', and the total complexity from n^2 to
n*log(n).

This patch makes that change, finding the overlapping extents using the
tree, and then using the existing search_queue code to check the locks
associated with each extent.

This has been tested on Saturn-p2 by Dennis, using explicit lock ahead
tests, and has also had general Lustre testing done there to verify no
regressions.

Bugs Affected: 837878
Bug Action: RESOLVE