Symlink ko2iblnd-probe to /usr/sbin
As a "quick and dirty fix" for bug 834643 we added a symlink for
ko2iblnd-probe to /usr/sbin where the ko2iblnd module is expecting
to find this utility. That was commit 3ee03d73 in pull-request #41.
In that commit we only added this symlink in the ari_s-rhine package.

We now have a 2.7 build on a gemini Pearl system, so we need the "fix"
there as well. Here we also add it to the ari_s package for
completeness.
Also updated the original fix to use appropriate obs macros.

Bugs Affected: 834643
Bug Action: RESOLVE