Use percpu counter and rwsem in lu_cache_shrink
The algorithm for counting freeable objects in the lu_cache shrinker
does not scale with the number of cpus. The LU_SS_LRU_LEN counter for
each cpu is read and summed at shrink time. This mod uses the kernel's
percpu counter library to maintain the lru_len count. With this
library, the summation occurs when a particular cpu's counter exceeds
a threshold. lu_cache_shrink_count() returns the last computed sum; in
other words, it returns an approximation of the number of objects
available for freeing.

This mod also replaces the lu_sites_guard mutex with a rw semaphore.
The lock protects the lu_site list, which is only modified when a file
system is mounted or unmounted. The mutex forces concurrent
lu_cache_shrink callers to wait. This can cause RCU stalls or heartbeat
faults on busy systems with little free memory.

Bugs Affected: 836565
Bug Action: RESOLVE
Bug Comment: Use percpu counter in lu_cache_shrink. Replace the
lu_sites_guard mutex with a rw semaphore.