Make drop_cache unconditionally discard read lock pages
A significant amount of time is sometimes spent during
drop_caches checking pages to see if they are covered by
another read lock.  Since all read locks will be destroyed
by (ldlm_ns_)drop_cache, this is a waste of time.

We do this by using the LDLM_FL_DISCARD_DATA flag, which is
currently used for special cases where you want to destroy
the dirty pages under a write lock rather than write them
out.

We set this flag on all the PR locks which are slated for
cancellation by ldlm_prepare_lru_list (when it is called
from ldlm_ns_drop_cache).

The case where another lock does cover those pages (and is
in use and so does not get cancelled) is safe for a few
reasons:

1. When discarding pages, we wait (discard_cb->cl_page_own)
until they are in the cached state before invalidating.
So if they are actively in use, we'll wait until that use
is done.

2. Removal of pages under a read lock is something that can
happen due to memory pressure, since these are VFS cache
pages.  If a client reads something which is then removed
from the cache and goes to read it again, this will simply
generate a new read request.

This has a performance cost for that reader, but if anyone
is calling drop_cache while actively doing I/O in that
namespace, then they cannot ask for good performance.

Basic testing shows this to be safe, and in the case of
using lock ahead to create 80,000 read locks on one
resource, then reading the whole file to populate the
pages, this improves cleanup time from ~60 seconds to
~ 0.5 seconds.  This speed improvement will only apply
to the case of many locks on a single resource, but 2.7
makes this possible with or without lock ahead.

Bugs Affected: 838530
Bug Action: RESOLVE