Use correct iterator to place page in array
The code "puts" an array of pages and then places the ones
with no references in another array to be freed.

Previously, I used the iterator for checking the first
array to place objects in the second.  But that placement
is conditional on pages not being in use, so if any pages
were in use, this would create a gap in the second array,
resulting in a panic.