LU-8010 mdt: fix orphan layout_lock cases for restore
Previously to this patch layout was not unlocked when a restore
was failed before being sent to CT, leading to a situation where
other requestors hang and also to an orphan restore_handle to be
kept on CDT's list of registered restore actions.
Only way to clear situation then, was to stop CDT.
This situation could at least occur if a restore was canceled
but the CT does not handle cancel operation, allowing the restore
to complete but also to have new restore requests to be
registered in the mean time and then to be failed due to
incompatible (no longer released) file state.

Also fix similar deadlock cases where layout lock was taken for
previously started restore requests upon CDT restart, by forcing
their replay/restart. This should also strengthen overall HSM
recovery process.

Lustre-commit: c13ddec8e1cd3a63c16e08f28749771200b92f1b
Lustre-change: http://review.whamcloud.com/19710

Signed-off-by: Bruno Faccini <bruno.faccini@intel.com>
Signed-off-by: Bob Glossman <bob.glossman@intel.com>
Change-Id: Ib1ba9156793a230d256ff80d74372813f10b0321
Reviewed-by: John L. Hammond <john.hammond@intel.com>
Reviewed-by: jacques-Charles Lafoucriere <jacques-charles.lafoucriere@cea.fr>
Reviewed-by: Aurelien Degremont <aurelien.degremont@cea.fr>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>
Reviewed-on: http://review.whamcloud.com/21047
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>