Use granted extent tree to update kms
Currently, ldlm_extent_shift_kms does a linear search of
the list of granted locks on the resource.  This is
problematic because 2.7 and lockahead make it possible to
have thousands of locks on a resource.

When cleaning up locks, we end up doing a linear search of
this list for every lock on the object - complexity n^2.

This was taking an unacceptable amount of time and causing
heartbeat losses.

This can be avoided by using the red black binary trees
which store the extents of granted locks, reducing the
complexity of an individual search from 'n' to 'log n',
and the total complexity from n^2 to n*log(n).

Note that there are multiple trees, one for each lock mode.
This necessitates using kms and old_kms, since we check
every tree, even though we check the kms against only one
lock per tree.

This has been tested on Saturn-p2 using the reproducer for
bug 837707, and has also had general Lustre testing done
there to verify no regressions.

Bugs Affected: 838695
Bug Action: RESOLVE