Always update lli_has_smd
Currently, in cl_file_inode_init, we only update
lli_has_smd for new inodes (this tells us whether or not
OST objects have been allocated yet).  However, when this
code is called for an existing inode, there is a case where
the OST objects have been allocated, and then we should
update lli_has_smd.  Since it's always safe to do that, we
update it unconditionally.

Here's the scenario in which this happens (courtesy of
Andrew Perepechko at Seagate):
1) the file is created without objects (O_CREAT|O_RDONLY)
and no lsm is returned to the client, cl_file_inode_init()
initializes lli_has_smd to 0
2) the file is opened for write (O_CREAT|O_WRONLY), objects
are allocated for the file and the lsm is returned to the
client, but no lli_has_smd reinitialization happens in
cl_file_inode_init(), because the cl_object has already been
allocated, no lli_has_smd reinitialization happens in
ll_prep_inode() because no lock is returned to the client,
so lli_has_smd stays equal to 0
3) the file is opened for read (O_RDONLY|O_DIRECT),
lli_has_smd stays equal to 0 for the same reason as above
4) the check in ll_direct_IO fails and -EBADF is returned

If the inode is reclaimed instanly after (1) as before the
ll_d_iput removal (Patch for CLSTR-5579/833608),
cl_file_inode_init() will allocate a cl_object and set
lli_has_smd to 1.

Bugs Affected: 839340
Bug Action: RESOLVE