Symlink ko2iblnd-probe to /usr/sbin
As a "quick and dirty fix" for bug 834643 we added a symlink for
ko2iblnd-probe to /usr/sbin where the ko2iblnd module is expecting
to find this utility. That was commit 3ee03d73 in pull-request #41.
In that commit we only added this symlink in the ari_s-rhine package.

We're now deploying a 2.7 client on eLogin, so we need the "fix"
there as well.

Bugs Affected: 840462
Bug Action: RESOLVE