Add modulefile to dmp_c build.
Adding the modulefile to the dmp_c build
will allow compilation against liblustreapi on elogin nodes.
This change will install the package config files and the
module file that was added to the ari_s-rhine builds.

Test Description:
I tested this change on percival-elogin3. I ran the rest script:
http://bugzilla.us.cray.com/attachment.cgi?id=20249
and it was able to statically and dynamically compile.

Bugs Affected: 788353
Bug Action: UPDATE