LU-7759 llite: handle inactive OSTs better in statfs
Change the order of checks for inactive OSCs in lov_prep_statfs_set()
so that administratively disabled OSTs do not generate any output in
"lfs df" at all, to avoid needlessly cluttering the output.

Enable the lazystatfs mount option by default, so that "df" does not
hang when an OST is temporarily offline.

Fix test-framework.sh to use "lfs df $MOUNT" instead of "df $MOUNT" to
determine when recovery is complete, now that "df" does not block.

This is a port patch from:
Lustre-commit: 23fde1f89bec0adf4f7181ccce5a236eac371a38
Lustre-change: http://review.whamcloud.com/19195

Signed-off-by: Andreas Dilger <andreas.dilger@intel.com>
Change-Id: I993761a7eb120a36a1b80c2822cb6d8011ccab07
Reviewed-by: John L. Hammond <john.hammond@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>
Reviewed-on: http://review.whamcloud.com/21034
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Dmitry Eremin <dmitry.eremin@intel.com>