841945 libcfs: Check if strtoul is already defined
On SLES 11 SP3 the OFED 3.18 compatibility layer includes a header that
defines strtoul to simple_strtoul. One of Lustre's headers does the same
thing causing a compile error. Modify the Lustre code to only make
the #define if strtoul is not already defined as a macro.

This change is not needed upstream.

Bugs Affected: 841945
Bug Action: RESOLVE