841829 kgnilnd: Remove read capability of cksum_test proc file
When the old create proc interface was deprecated cksum_test was updated
to use the new file operations table. Inadvertantly read was left as a capability
without actually defining a function that the file would use when someone tried
to read the file. This causes a kernel crash when cksum_test is read, though it
can only be done by the root user.

The fix is to remove the .read op from the fops table for the cksum_test
proc entry

Bugs Affected: 841829
Bug Action: RESOLVE