LU-7828 statahead: set sai_index_wait with lli_sa_lock held
It is the sponsor thread of the statahead thread that updates the
sai::sai_index_wait. Originally, it didn't hold the lli_sa_lock when it
did that. Because of out-of-order execution others may miss to wakeup
such thread.

On the other hand, if the statahead RPC gets failure, it
should wakeup the sponsor thread, not the statahead thread.

Bugs Affected: 837406
Bug Action: RESOLVE
Bug Comment: