LU-8494 ldlm: Return -ERESTARTSYS from wait
In ldlm_completion_ast, when we sleep waiting for a lock
to be granted or cancelled, we sleep interruptibly. When
interrupted by a signal, we return -EINTR. This prevents
syscalls from being restarted if caught here.

Change that -EINTR to an -ERESTARTSYS.

Tested with a simplified reproducer, confirmed fix.

Bugs Affected: 840648
Bug Action: RESOLVE