Bz839670: debug patch to count users of loh_ref
This mod is intended to be temporary. It should never be
included in a release branch. It is a debug patch that counts
the number of active uses of loh_ref divided by user type. A
count for the type is incremented on get/decremented on put.
The patch is intended to help find what is not releasing the
ref count.

Bugs Affected: 839670
Bug Action: UPDATE
Bug Comment: Adding a debug patch to help find location of
missing reference count put. For internal use only.