LU-4533 lprocfs: Increase OBD_HIST_MAX to 256
Increase OBD_HIST_MAX to 256 to match OSC_MAX_RIF_MAX so
that the rpcs in flight histogram which is displayed by
reading the rpc_stats proc interface is correct.

Signed-off-by: Ryan Haasken <haasken@cray.com>
Change-Id: I6118bf258a76e7791071744f9862d018b732672e

Note the memory cost of this is:
sizeof(unsigned long)*(OBD_HIST_MAX_new -
OBD_HIST_MAX_old)*OST_count

For a 64 bit system with 2048 OSTs, this is:

8 bytes*256-32*2048 = 3.5 MiB

Which is trivial.

Bugs Affected: 799412
Bug Action: RESOLVE