LU-8418 libcfs: do not call empty lnet upcall
An attempt to call lnet upcall in lbug_with_loc
may cause lbug thread to hang even if lnet upcall
is not set or set to non-existing path.
The patch allows to skip lnet upcall if it is set
to "NONE" keyword.

Change-Id: I20af6874116542d16bcc9a9eb75c813a124e346d
Seagate-bug-id: MRP-2939
Signed-off-by: Alexander Zarochentsev
<alexander.zarochentsev@seagate.com>

Note this patch changes the default to 'NONE', so the
upcall is disabled by default.  This would break any
existing users, but we are not aware of any.

Bugs Affected: 842730
Bug Action: RESOLVE