MRP-3720 ptlrpc: update req timeout if resending happened
When the server drops duplicate request processing, the client and
the server have different deadline for the same request. The server
operates with the first copy and the client operates with the second.

This patch adds request deadline updates if a duplicate request is
found.

A fix for MRP-3582 changed lock callback prolong calculation to use
request deadline in case when service estimate changed since the
request has beed created. Using outdated deadline may cause
insufficient prolong timeout and subsequent client eviction.

Signed-off-by: Alexander Boyko <alexander.boyko@seagate.com>
Signed-off-by: Vladimir Saveliev <vladimir.saveliev@seagate.com>
Change-Id: Ic1c92d122760dce1d2603c29af9b6720ecd8424a
Reviewed-on: http://es-gerrit.xyus.xyratex.com:8080/12939
Reviewed-by: Vitaly Fertman <vitaly.fertman@seagate.com>
Tested-by: Elena V. Gryaznova <elena.gryaznova@seagate.com>