LU-5794 osd: additional checks to verify credits calculation
every operation counts credits consumed, so we can check when
credits are exceeded. now the number of declarations doesn't
matter for the checks, it's only credits. i.e. it's OK to
declare a single write for a specific range and then make many
writes to that range within a single transaction.

the patch fixes few issues on osd-zfs making it impossible to
declare change to an object being created in a same transaction.

Lustre-commit: 92d28a7f35bcdc718e2bdcec48bcdeacd0e8f4fe
Lustre-change: http://review.whamcloud.com/12169

Signed-off-by: Alex Zhuravlev <alexey.zhuravlev@intel.com>
Change-Id: I427e2a3b96623cdc23e5d7734f8ecc7cc16382bd
Reviewed-on: http://review.whamcloud.com/22516
Tested-by: Jenkins
Reviewed-by: Fan Yong <fan.yong@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Lai Siyao <lai.siyao@intel.com>
Reviewed-by: John L. Hammond <john.hammond@intel.com>

- Conflicts in lustre/osd-ldiskfs/osd_internal.h
    - Trivial conflict with Cray commit 69e503f256a3c92c59d64287fd72eaa78c16bf6c
    - Trivial conflict with Cray commit d74e1ded00d3d8dbbfc2a1ac4a40b84f8c5e3ed0