MRP-3904 osd-ldiskfs: cannot inherit the setgid from parent directory
An earlier commit "osd-ldiskfs: pass uid/gid/xtime directly to ldiskfs"
uncovered a bug in osd-ldiskfs.

According to the SGID semantics, if a file is created in a SGID directory,
it inherits the group of the directory. If a directory is created in a
SGID directory, it also inherits the SGID bit.

mdd_create_sanity_check() updates client-provided attributes so
that they follow these rules. However, osd_mkdir() clears the
SGID bit so that this bit does not get into the backend fs.

Things worked fine until the forementioned commit because the
backend fs itself contains this logic. However, we cannot
assume that the backend fs should perform this job, so this
patch makes ldiskfs take into account the SGID bit provided
by the upper layers.

Change-Id: Id461fef3413ee48368b4b764cc6ffe853135153c
Signed-off-by: Andrew Perepechko <andrew.perepechko@seagate.com>
Reviewed-on: http://es-gerrit.xyus.xyratex.com:8080/12452
Reviewed-by: Andriy Skulysh <andriy.skulysh@seagate.com>
Reviewed-by: Alexander Zarochentsev <alexander.zarochentsev@seagate.com>
Tested-by: Elena V. Gryaznova <elena.gryaznova@seagate.com>