LU-7885 build: Add option for extra Module.symvers
This change updates our existing code to match the latest patchset from
LU-7885. The description of the LU-7885 patch under review upstream is
as follows:

Lustre would not install properly on a particular platform because it
wasn't built with the gni Module.symvers. While this only causes a
warning on SLES, it stops the installation on CentOS.

Add option to build with an out-of-tree Module.symvers.

Signed-off-by: Chris Horn <hornc@cray.com>
Change-Id: Icc6977f913ab31d69fdbb86f2e2bc475c6c8a853
Reviewed-on: http://review.whamcloud.com/19009