Limit the maximum size of the ldlm_lock lru
When lru resizing is enabled, the number of locks in the lru is limited only
by the server_lock_volume (slv) set by the server. Typical slv values do not
impose much of a limit on the lru size in practice. Lock counts are often hit
tens and even hundreds of thousands. Neither the ldlm pool shrinker nor drop
caches can cancel these large number of locks effectively. This mod introduces
a default cap on the maximum size of the LRU and adds a module parameter to
override the default.

The module parameter is: options ptlrpc ldlm_max_lru_size=<value>
where value is a nonnegative integer indicating the max number of locks in the
lru. If value == 0, then no limit is imposed (i.e. same behavior as before
this mod).

This mod is intended as a short term fix until the slv logic is reworked
upstream (LU-7266).
Bugs Affected: 839445
Bug Action: RESOLVE
Bug Comment: