Remove post build checks from lnet-devel
The lnet-devel build requires BuildConflicts
post-build-checks to build against the sles12sp2 kernel. Otherwise
the configure step fails. This is because uname detects the
kernel version of the kernel that is installed on the build machine
not installed in the build root. post-build-checks installs a script
to work around this however the script fails in this instance. Adding
BuildConflicts post-build-checks prevents the installation of
post-build-checks.

Bugs Affected: 845139 839015
Bug Action: RESOLVE