Do not set S_NOSEC for existing inodes in ll_iget
The original LU-8656 patch attempted to set S_NOSEC for
both new and pre-existing inodes in ll_iget.

This requires taking the i_mutex in ll_iget for existing
inodes.  This has the effect of serializing many lookups
which were previously parallel, with a disastrous impact
on metadata performance.

Setting S_NOSEC on these inodes in ll_iget is an
optimization to prevent taking the i_mutex for the first
write to a file, but it is not necessary in general to
keep the 'lockless write' behavior of Lustre.

The first write to the file will take the i_mutex, then set
the S_NOSEC flag (in file_remove_suid) during the write.

This means the first parallel writes to a file can be
briefly delayed, but the problem does not persist.  Quick
testing shows this to still allow good write performance.

Bugs Affected: 845659
Bug Action: RESOLVE