LDEV-507 mgc: use remote llog under read only mode
Under read only mode, we ignore to sync the server
(MDT or OST) local llog copy with the remote llog,
so the it may has no local llog copy or its local
llog copy is incomplete. Under such case, we should
use the remote llog directly when start the server.

Signed-off-by: Fan Yong <fan.yong@intel.com>
Change-Id: Ia9bec22f6dbbc2214ff0990c86fdfbcc722ab1c9
Reviewed-on: https://review.whamcloud.com/22582
Tested-by: Jenkins
Reviewed-by: Andreas Dilger <andreas.dilger@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: John L. Hammond <john.hammond@intel.com>