845778 build: Handled unpackaged file error
This is a cherry-pick of feature-2.7 commit
e6e6386b67ac7364b057ca78408cabd68cd140e4 to Cray-master. Cray-master
does not have the perm.conf file, so we're just taking the "light
refactoring" changes mentioned in the original commit.

The merge of b_neo_stable introduced a new file: /etc/lustre/perm.conf
We do need/want to include this file in our packages so it should be
excluded, however some of our packages do not currently exclude it.

A little light refactoring is included here so we're not just
deleting files from the build-root, but instead using the %exclude
rpm macro.

Bugs Affected: 845778
Bug Action: RESOLVE

 Conflicts:
	cray-lustre-dmp_s.spec
	cray-lustre-gem_c.spec
	cray-lustre-gem_s-cos.spec
	cray-lustre-gem_s.spec