845778 build: Handled unpackaged file error
The merge of b_neo_stable introduced a new file: /etc/lustre/perm.conf
We do need/want to include this file in our packages so it should be
excluded, however some of our packages do not currently exclude it.

A little light refactoring is included here so we're not just
deleting files from the build-root, but instead using the %exclude
rpm macro.

Bugs Affected: 845778
Bug Action: RESOLVE