LU-5835 ptlrpc: Introduce iovec to bulk descriptor
Added a union to ptlrpc_bulk_desc for KVEC and KIOV buffers.
bd_type changed to be a bit mask.
Bits are set in bd_type to specify {put,get}{source,sink}{kvec,kiov}
changed all instances in the code to access the union properly
ASSUMPTION: all current code only works with KIOV and new DNE code
to be added will introduce a different code path that uses IOVEC

As part of the implementation buffer operations are added
as callbacks to be defined by users of ptlrpc.  This enables
buffer users to define how to allocate and release buffers.

Lustre-commit: 7cb2e4b642b51b42f9f3a448251774929073e17c
Lustre-change: http://review.whamcloud.com/12525

Signed-off-by: Amir Shehata <amir.shehata@intel.com>
Change-Id: I86d718a668e7c0b422b96b393a6a69e5b8e89291
Reviewed-by: wang di <di.wang@intel.com>
Reviewed-by: Liang Zhen <liang.zhen@intel.com>