Fix infinite loop in test15
There is undefined behavior in test15 (integer overflow in
a multiply), which GCC optimizations turn in to an infinite
loop. This only happens in newer GCC versions, like on
CentOS 7.

This fix tells GCC some of the operands in the
multiplication are unsigned longs, so it fixes the
integer overflow (start is a 64 bit int, but i and
unspecified constants are 32 bit ints, so an overflow
occurs in the multipication).

Because the multiplication is by a constant, GCC multiplied
i by that constant, and used that as the loop condition.

The overflow which results causes the aforementioned
infinite loop. This change clarifies intent, removing the
undefined behavior. (Presumably forcing GCC to either not
modify the loop condition or use 64 bit storage for it.)

With this change, the test completes normally.

Bugs Affected: 845997
Bug Action: RESOLVE