LU-8284 osd-ldiskfs: i_size updates from BRW should be atomic
In osd_write_commit(), it is possible that somewhere between
i_size_read() check and i_size_write() another write updates
i_size to a larger value. In the end, we will have a lower
i_size value than we should.

Change-Id: I720a6cc1214e3e6b32df4f9e2c1017a47e7962f0
Signed-off-by: Andrew Perepechko <andrew.perepechko@seagate.com>
Reviewed-by: Alexander Zarochentsev <alexander.zarochentsev@seagate.com>
Reviewed-by: Alexander Nikolaevich Boyko <alexander.boyko@seagate.com>
Seagate-bug-id: MRP-3528
Reviewed-on: http://review.whamcloud.com/20816
Reviewed-by: Andreas Dilger <andreas.dilger@intel.com>
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Alexander Boyko <alexander.boyko@seagate.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>
---
This bug results in file corruption, where part of the file
which as written is not readable because it is beyond the
reported file size.  This is a server side problem.

Bugs Affected: 846085
Bug Action: UPDATE