Do not set request_only with O_NONBLOCK
If request only is set on a file descriptor opened with
O_NONBLOCK, the ldlm code recognizes the resulting
requests as lock ahead requests.  This means lock requests
get true nonblocking behavior, which means IO requests will
fail with -EAGAIN if there is a blocking lock.

This is unacceptable, so if a file is opened O_NONBLOCK, we
ignore the request only setting.  This is OK since request
only is advice, designed to enhance performance in special
cases whereas O_NONBLOCK is a fundamental part of POSIX
semantics and must be respected.

Bugs Affected: 846437
Bug Action: UPDATE