Correct handling of PR locks when updating kms
The fix for 838695 switched to using the interval tree when
updating known minimum size (kms) on lock cancellation.
However, it assumed that the highest interval in the tree
had the highest 'end'. Since the tree is sorted by the
start of the intervals, and PR (read) locks can overlap,
this is not valid for those locks.

So instead of always stopping at the first PR lock, we must
continue until either we find a lock with a higher end than
our old kms, or we check all of the locks.
Here's the relevant text from the upstream commit,
explaining why this still isn't too bad:

Since the tree is sorted by 'start' and PR locks can
overlap, we cannot easily use the tree to find the PR lock
with the 'highest end'. But we can still dramatically
improve the average performance by walking the tree from
the highest starting lock, since locks with a high starting
point will more often have high end points.

Note that we test in_max_high for each tree, and if kms is
in_max_high, we skip that tree. This is safe because the
kms_ignore flag means that kms from that tree can be lower
than in_max_high for that tree, but it can never be higher,
since in_max_high includes locks with the kms_ignore flag
set.

Bugs Affected: 846380
Bug Action: RESOLVE