LU-6971 cleanup: not support remote client anymore
There are several obsolete sub commands for lfs to work with
remote client. We do not support that anymore, and should be
deleted along with any kernel code related to remote client.

Partial backport to 2.7. We need to remove some calls to
getgrnam and getpwnam which cause statically linked binaries
to throw compiler warnings (and panic if not executed in the right
environment). Since this issue is with glibc and goes back decades
simply removing the offending, unused code from liblustreapi
is a simple fix.

Signed-off-by: Ben Evans <jevans@cray.com>

Bugs Affected: 835858
Bug Action: RESOLVE