847301 build: Remove long option from git describe
With our current w.x.y.z tags on stable-2.7 the use of the long
option results in a 5th version number, e.g. 2.7.1.13.0. The other
benefits it provides are superfluous as the git commit hash will be
included in both the package name, via the Release field in our spec
files, and the internal build version via other means, and the knowledge
that we're zero commits from a tag is arguably worthless.

Bugs Affected: 847301
Bug Action: UPDATE