847301 build: Only match tags with '_cray'
This is more of a defensive thing to make sure the tag based versioning
doesn't match non-cray created tags that may get pushed to our
repository.

While we only want to match tags with the '_cray' string, that string
should not appear in the _tag or _version variables.

Bugs Affected: 847301
Bug Action: UPDATE