808203 build: Update location of headers in lnet-devel package
The change in 5566be44ce is for rhine only builds, but it has the side
effect of changing the pkg-config flags passed to builds of DVS that
use the existing/old lnet-devel package. The pkg-config flags contain
an include directive for lustre's config.h file, but the location of
the file has changed with the rhine devel package.
Additionally, some include paths were removed from the pkg-config
flags so we want to move the location of headers in the devel package
to match the current pkg-config flags.

Bugs Affected: 808203
Bug Action: UPDATE