Remove extra MMSPACE_OPEN call
A merge error in feature-2.7 commit
780a3ee447203da3d2f84b9baa50e09ae957a270
left a call to MMSPACE_OPEN in cfs_tracefile_dump_all_pages

Because the MMSPACE_OPEN call was replaced by direct calls
to the get_fs/set_fs functions, leaving it in caused
userspace processes which called this function to keep the
kernel addr limit on exit from the function, rather than
restoring the userspace address limit.

The next time such a process tried to read or write data,
this incorrect address limit was used to decide whether to
call memcpy or copy_from/to_user.  The kernel address limit
resulted in calling memcpy, which then caused a fatal
"unable to handle kernel paging request" when it was called
with a userspace pointer.

The fix is just to remove this MMSPACE_OPEN call.

Bugs Affected: 848346
Bug Action: RESOLVE