SEA-327 tests: SKIP sanity/300[e-i] for older MDS version
This is a partial port of patch for LU-6984.
"LU-6984 Failure to delete over a million files
in a DNE2 directory."

The test sanity/300[e-i] needs to be SKIPPed for older
version of MDS esp in cases where
"LU-3531 mdt: delete striped directory" isn't present OR
a supported feature.

Test-Parameters: trivial testlist=sanity
Signed-off-by: Parinay Kondekar <Parinay.Kondekar@seagate.com>
Change-Id: Ie07b340a1296e4b5b7889010caf2b4dd622d2c32
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Jian Yu <jian.yu@intel.com>
Reviewed-by: John L. Hammond <john.hammond@intel.com>