Correct CEF_REQ_ONLY flag value
Commit d54e7efdccad05f9f8a1eac7ebd4f0088e3d82f5, LU-6271,
added a new osc lock flag, CEF_LOCK_MATCH.  This has the
same value as the CEF_REQ_ONLY flag for lockahead.

This flag is only used in memory on the local node, so we
simply move it to a different value.  (Note this is the
value used by the equivalent flag in upstream master.)

Bugs Affected: 849166
Bug Action: RESOLVE