CentOS 6.7 (573.26.2) support
ext4-large-dir.patch is required for our version of the
pdirop patch (see commit
51cc2f4323cdcf9172dc63062932c21fab4ea70f).
ext4-use-correct-inode is required for the rhel6 version
of the large-dir patch. (It is already in our 6.5 and 6.6
series.)

Note this is NOT intended to be production support for
servers on this kernel, which we expect to get from Intel/
Seagate in a future drop of NEO. This is build support to
aid internal testing.

Bugs Affected: 849505
Bug Action: RESOLVE