Add lustre include files to compute nodes
I modified the compute spec files so that the compute rpms now include
the header files. To do this I just took the necessary code from the
service spec files. Since we are moving away from these builds with
the meta spec builds I didnt feel like it was worth it to develope
a whole new solution to the problem.

To test the builds I built the packages on lusbld01 and compared their
contents to the service builds.

Bugs Affected: 849651
Bug Action: RESOLVE