Correct rhel7.3 patch series
When we brought in the final round of b_neo_stable patches
to feature-2.7, I merged the rhel7.3 ldiskfs series file,
and made a mistake.  The patch series added two new
rhel7.3-specific patch files, specifically:
+rhel7.3/ext4-osd-iop-common.patch
+rhel7.3/ext4-pdirop.patch

We need those were using before.

In a slight twist, we need the upstream rhel7 version of
one patch:
+rhel7/ext4-osd-iop-common.patch

And our rhel7.2 version of the other patch:
+rhel7.2/ext4-pdirop.patch

This is because the 7.3 kernel for Athena is not *quite* as
new as upstream expects.  It's .2.2, not .10.2.  So we
can't use the 7.3 or master rhel7 version, which are aimed
at (respectively) an earlier and a later kernel version.

There do not appear to be any fixes in the diff between
our versions and the rhel7.3 and master versions of the
ext4-pdirop patch.

Bugs Affected: 850551
Bug Action: RESOLVE